// server/src/user-fav-reviews/dto/types/author-likes-raw.ts
export type AuthorLikesRaw = {
  reviewAuthorId: string;
  reviewAuthorNickname: string;
  reviewAuthorAvatar: string;
  reviewTitle: string;
  likerUserId: string;
  likerNickname: string;
  likerAvatar: string;
  releaseId: string;
  releaseTitle: string;
  createdAt: Date;
  releaseImg: string;
};